requirements.txt
python-dotenv==0.20.0
selenium==4.1.3
setuptools==62.1.0
beautifulsoup4==4.11.1